# [numpy] FFI binary bitwise ops 